rename rpc module to muxrpc
