+++ src/actors/mod.rs
@@ -1,5 +1,5 @@
 pub mod ctrlc;
 pub mod jsonrpc;
+pub mod muxrpc;
 pub mod network;
 pub mod replication;
-pub mod rpc;
